New code to be cleaned with float and double

+++ src/_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   _utils.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pmarkaid <pmarkaid@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: dbejar-s <dbejar-s@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 14:05:47 by pmarkaid          #+#    #+#             */
-/*   Updated: 2024/11/24 12:39:35 by pmarkaid         ###   ########.fr       */
+/*   Updated: 2024/11/25 14:13:14 by dbejar-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,8 @@ void	print_map_struct(t_map *map)
 	printf("\nCeiling Colors (RGB):\n");
 	printf("  C: %d, %d, %d\n", map->c[0], map->c[1], map->c[2]);
 	printf("\nMap Dimensions:\n");
-	printf("  Width: %zu\n", map->w_map);
-	printf("  Height: %zu\n", map->h_map);
+	printf("  Width: %d\n", map->w_map);
+	printf("  Height: %d\n", map->h_map);
 	printf("\nMap Layout:\n");
 	i = 0;
 	while (map->grid[i])
@@ -40,8 +40,8 @@ void	print_map_struct(t_map *map)
 
 void	print_arrays(t_macro *m, char **visited)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 
 	ft_printf(1, "Grid:\n");
 	for (i = 0; i < m->map->h_map; i++)
